+++ finalchat/resource.h
@@ -14,8 +14,11 @@
 #define IDC_EDIT_PASSWORD               1002
 #define IDC_BUTTON_SIGNIN               1003
 #define IDC_BUTTON_SIGNUP               1004
-#define IDC_LIST1                       1005
 #define IDC_LIST_USERNAME               1005
+#define IDC_LIST_MSG                    1006
+#define IDC_EDIT_MSG                    1007
+#define IDC_BUTTON1                     1008
+#define IDC_BUTTON_SEND                 1008
 
 // Next default values for new objects
 // 
@@ -23,7 +26,7 @@
 #ifndef APSTUDIO_READONLY_SYMBOLS
 #define _APS_NEXT_RESOURCE_VALUE        130
 #define _APS_NEXT_COMMAND_VALUE         32771
-#define _APS_NEXT_CONTROL_VALUE         1006
+#define _APS_NEXT_CONTROL_VALUE         1009
 #define _APS_NEXT_SYMED_VALUE           101
 #endif
 #endif
