+++ server/ServerSocket.h
@@ -2,7 +2,15 @@
 #include <afxsock.h>
 #include "serverDlg.h"
 #include <thread>
+#include <fstream>
+#include <string>
 class CserverDlg;
+
+struct SendMsgStruct {
+	CString targetName;
+	CString yourName;
+	CString msg;
+};
 
 class ServerSocket : public CAsyncSocket
 {
@@ -12,6 +20,14 @@ public:
 	~ServerSocket();
 	virtual void OnAccept(int nErrorCode);
 	virtual void OnReceive(int nErrorCode);	
-	void ReceiveThread();
+	CString GetSignal(CString msg);
+	bool GetUserPass(CString msg);
+	bool CheckIfUsernameLogin(CString username);
+	CString tempName;
+	std::ifstream fileIn;
+	std::ofstream fileOut;
+	SendMsgStruct GetSendMsgStruct(CString msg);
+	bool SignUpUser(CString msg);
+	CString GetUsernameLogout(CString msg);
 };
