+++ server/taikhoan.txt
@@ -2,3 +2,7 @@ taikhoan
 123
 tk1
 123
+tk2
+123
+tk4
+123
