+++ x64/Release/taikhoan.txt
@@ -1,0 +1,1 @@
+aaa bbb
